(refactor) Move generateCoordinates to a new file
